# static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "XM/Exness Ready Linux MetaTrader 5 API",
    "description": "REST API to manage multiple MetaTrader 5 accounts running inside Docker/Wine.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/primary": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Primary Data",
        "operationId": "primary_data_api_primary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/data": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "All Data",
        "operationId": "all_data_api_data_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/accounts": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "List Accounts",
        "operationId": "list_accounts_api_accounts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Create Account",
        "operationId": "create_account_api_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/accounts/{acc_id}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Get Account",
        "operationId": "get_account_api_accounts__acc_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Acc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Accounts"
        ],
        "summary": "Update Account",
        "operationId": "update_account_api_accounts__acc_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Acc Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Accounts"
        ],
        "summary": "Delete Account",
        "operationId": "delete_account_api_accounts__acc_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Acc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{acc_id}/start": {
      "post": {
        "tags": [
          "Lifecycle"
        ],
        "summary": "Start Account",
        "operationId": "start_account_api_accounts__acc_id__start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Acc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{acc_id}/stop": {
      "post": {
        "tags": [
          "Lifecycle"
        ],
        "summary": "Stop Account",
        "operationId": "stop_account_api_accounts__acc_id__stop_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Acc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{acc_id}/restart": {
      "post": {
        "tags": [
          "Lifecycle"
        ],
        "summary": "Restart Account",
        "operationId": "restart_account_api_accounts__acc_id__restart_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Acc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{acc_id}/data": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Account Data",
        "operationId": "account_data_api_accounts__acc_id__data_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Acc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{acc_id}/price": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Account Price",
        "operationId": "account_price_api_accounts__acc_id__price_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Acc Id"
            }
          },
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{acc_id}/bars/{timeframe}": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Account Bars Timeframe",
        "operationId": "account_bars_timeframe_api_accounts__acc_id__bars__timeframe__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Acc Id"
            }
          },
          {
            "name": "timeframe",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Timeframe"
            }
          },
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 100,
              "title": "Count"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{acc_id}/bars/{symbol}/{timeframe}": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Account Bars Symbol Timeframe",
        "operationId": "account_bars_symbol_timeframe_api_accounts__acc_id__bars__symbol___timeframe__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Acc Id"
            }
          },
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "timeframe",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Timeframe"
            }
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 100,
              "title": "Count"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{acc_id}/positions": {
      "get": {
        "tags": [
          "Trading"
        ],
        "summary": "Account Positions",
        "operationId": "account_positions_api_accounts__acc_id__positions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Acc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{acc_id}/orders": {
      "get": {
        "tags": [
          "Trading"
        ],
        "summary": "Account Orders",
        "operationId": "account_orders_api_accounts__acc_id__orders_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Acc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Trading"
        ],
        "summary": "Create Order",
        "operationId": "create_order_api_accounts__acc_id__orders_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Acc Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{acc_id}/history": {
      "get": {
        "tags": [
          "Trading"
        ],
        "summary": "Account History",
        "operationId": "account_history_api_accounts__acc_id__history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Acc Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "example": "XM Account"
          },
          "login": {
            "type": "string",
            "title": "Login",
            "example": "168892363"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "example": "Essa123$$$"
          },
          "server": {
            "type": "string",
            "title": "Server",
            "example": "XMGlobal-MT5 2"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol",
            "default": "XAUUSDm",
            "example": "GOLD"
          }
        },
        "type": "object",
        "required": [
          "name",
          "login",
          "password",
          "server"
        ],
        "title": "AccountCreate"
      },
      "AccountUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Login"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "server": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "type": "object",
        "title": "AccountUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}